Add vcr support for scenarios
Replace aruba-supplied steps in scenario with customised step, allow us
to wrap the executed code in vcr configuration and cassette management
code. This seems quite novel as I couldn't find an example of mixing
vcr and aruba this way.

Includes the vcr cassette for the scenario. This seemed sufficiently
different to prior recordings for the specs (primarily the content being
formatted differently) that I decided to include this new recording.

Also simplifies the step relating to environment variable which was
needlessly parameterised, inviting failures in future if a different
variable name was used in future scenarios.