Add scenario for caching
The step def for verifying the number of requests made checks for output
logged to STDERR. This is a bit of a hack and I'd ideally make use of
webmock. At a later date I'll probably replace aruba with in-process
execution of the the code. This should also speed up running the
scenarios.